refactor: ApplicationProcessQueryService 패키지 변경
